move font into static directory
